Move download to own file
